feat(desktop): add Cybara ACP harness preset
